Move language types to different package
